Move YamlConnectionFactory to not be a nested class
