+++ yaml-tests/src/main/java/com/apple/foundationdb/relational/yamltests/configs/MultiServerConfig.java
@@ -21,7 +21,7 @@
 package com.apple.foundationdb.relational.yamltests.configs;
 
 import com.apple.foundationdb.relational.yamltests.MultiServerConnectionFactory;
-import com.apple.foundationdb.relational.yamltests.YamlRunner;
+import com.apple.foundationdb.relational.yamltests.YamlConnectionFactory;
 import com.apple.foundationdb.relational.yamltests.server.ExternalServer;
 
 import javax.annotation.Nonnull;
@@ -52,7 +52,7 @@ public class MultiServerConfig extends JDBCInProcessConfig {
     }
 
     @Override
-    public YamlRunner.YamlConnectionFactory createConnectionFactory() {
+    public YamlConnectionFactory createConnectionFactory() {
         return new MultiServerConnectionFactory(
                 MultiServerConnectionFactory.ConnectionSelectionPolicy.ALTERNATE,
                 initialConnection,
@@ -60,8 +60,8 @@ public class MultiServerConfig extends JDBCInProcessConfig {
                 List.of(createExternalServerConnection()));
     }
 
-    YamlRunner.YamlConnectionFactory createExternalServerConnection() {
-        return new YamlRunner.YamlConnectionFactory() {
+    YamlConnectionFactory createExternalServerConnection() {
+        return new YamlConnectionFactory() {
             @Override
             public Connection getNewConnection(@Nonnull URI connectPath) throws SQLException {
                 String uriStr = connectPath.toString().replaceFirst("embed:", "relational://localhost:" + externalServer.getPort());
